building out the cart within Checkout
Now I think about it, haveing the scan as part of the Checkout could be better.

Ideally, you would have a 'Cart' class, which items accumulate into,
then the checkout happens, and any final price adjustments happen there.
But, that would not show reduced prices in the cart. So the promo-rules
would be another class applied for both?